core/requirements.txt
faiss-cpu==1.8.0.post1
numpy==1.26.4
torch==2.2.2
transformers==4.44.2
nbformat==5.10.4
pinecone==5.1.0
python-dotenv==1.0.1
anthropic==0.34.2